+++ lib/flop.ex
@@ -1088,6 +1088,37 @@ defmodule Flop do
   end
 
   @doc """
+  Sets the page value of a `Flop` struct while also removing/converting
+  pagination parameters for other pagination types.
+
+      iex> set_page(%Flop{page: 2, page_size: 10}, 6)
+      %Flop{page: 6, page_size: 10}
+
+      iex> set_page(%Flop{limit: 10, offset: 20}, 8)
+      %Flop{limit: nil, offset: nil, page: 8, page_size: 10}
+
+  The page number will not be allowed to go below 1.
+
+      iex> set_page(%Flop{}, -5)
+      %Flop{page: 1}
+  """
+  @doc since: "0.12.0"
+  @spec set_page(Flop.t(), pos_integer) :: Flop.t()
+  def set_page(%Flop{} = flop, page) do
+    %{
+      flop
+      | after: nil,
+        before: nil,
+        first: nil,
+        last: nil,
+        limit: nil,
+        offset: nil,
+        page_size: flop.page_size || flop.limit || flop.first || flop.last,
+        page: max(page, 1)
+    }
+  end
+
+  @doc """
   Updates the `order_by` and `order_directions` values of a `Flop` struct.
 
   - If the field is not in the current `order_by` value, it will be prepended to
@@ -1124,8 +1155,19 @@ defmodule Flop do
       [:age, :species, :name]
       iex> flop.order_directions
       [:asc, :asc, :asc]
+
+  If a string is passed as the second argument, it will be converted to an atom
+  using `String.to_existing_atom/1`. If the atom does not exist, the `Flop`
+  struct will be returned unchanged.
+
+      iex> flop = push_order(%Flop{}, "name")
+      iex> flop.order_by
+      [:name]
+      iex> flop = push_order(%Flop{}, "this_atom_does_not_exist")
+      iex> flop.order_by
+      nil
   """
-  @spec push_order(Flop.t(), atom) :: Flop.t()
+  @spec push_order(Flop.t(), atom | String.t()) :: Flop.t()
   @doc since: "0.10.0"
   def push_order(
         %Flop{order_by: order_by, order_directions: order_directions} = flop,
@@ -1146,6 +1188,12 @@ defmodule Flop do
       )
 
     %{flop | order_by: order_by, order_directions: order_directions}
+  end
+
+  def push_order(flop, field) when is_binary(field) do
+    push_order(flop, String.to_existing_atom(field))
+  rescue
+    _e in ArgumentError -> flop
   end
 
   defp get_index(nil, _field), do: nil