+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [0.12.0] - 2021-08-11
+
 ### Added
 
 - Allow to define join fields in `Flop.Schema`.
@@ -10,12 +12,14 @@
 - Support filtering by compound fields.
 - New filter operator `empty`.
 - New filter operator `not_empty`.
+- New function `Flop.set_page/2`.
 
 ### Changed
 
 - Rename option `get_cursor_value_func` to `cursor_value_func`.
 - Silently ignore filters with `nil` value for the field or the value instead of
   raising an `ArgumentError`.
+- Allow passing a string as the second argument to `Flop.push_order/2`.
 
 ## [0.11.0] - 2021-06-13